fixing indentation bug in alttextify, updating output file handling in surface_alt_text, and starting to add evaluation on matplotlib example notebooks
